+++ gruntfile.js
@@ -43,14 +43,27 @@ module.exports = function (grunt) {
     eslint: {
       target: [
         'index.js',
-        'gruntfile.js'
+        'gruntfile.js',
+        'test/*.js'
       ]
+    },
+    simplemocha: {
+      options: {
+        ui: 'bdd',
+        reporter: 'dot'
+      },
+      unit: {
+        src: [
+          'test/*.spec.js'
+        ]
+      }
     }
   })
 
   require('load-grunt-tasks')(grunt)
 
-  grunt.registerTask('default', ['eslint'])
+  grunt.registerTask('test', ['simplemocha'])
+  grunt.registerTask('default', ['eslint', 'test'])
 
   grunt.registerTask('release', 'Bump the version and publish to NPM.', function (type) {
     grunt.task.run([